refactor: extract safari-nomodule-fix to a standalone plugin
