+++ packages/@vue/cli-service/lib/webpack/ModernModePlugin.js
@@ -1,26 +1,10 @@
 const fs = require('fs-extra')
 const path = require('path')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
-
-const { semver } = require('@vue/cli-shared-utils')
-const { projectModuleTargets } = require('../util/targets')
-
-const minSafariVersion = projectModuleTargets.safari
-const minIOSVersion = projectModuleTargets.ios
-const supportsSafari10 =
-  (minSafariVersion && semver.lt(semver.coerce(minSafariVersion), '11.0.0')) ||
-  (minIOSVersion && semver.lt(semver.coerce(minIOSVersion), '11.0.0'))
-const needsSafariFix = supportsSafari10
-
-// https://gist.github.com/samthor/64b114e4a4f539915a95b91ffd340acc
-const safariFix = `!function(){var e=document,t=e.createElement("script");if(!("noModule"in t)&&"onbeforeload"in t){var n=!1;e.addEventListener("beforeload",function(e){if(e.target===t)n=!0;else if(!e.target.hasAttribute("nomodule")||!n)return;e.preventDefault()},!0),t.type="module",t.src=".",e.head.appendChild(t),t.remove()}}();`
-
 class ModernModePlugin {
-  constructor ({ targetDir, isModernBuild, unsafeInline, jsDirectory }) {
+  constructor ({ targetDir, isModernBuild }) {
     this.targetDir = targetDir
     this.isModernBuild = isModernBuild
-    this.unsafeInline = unsafeInline
-    this.jsDirectory = jsDirectory
   }
 
   apply (compiler) {
@@ -87,36 +71,6 @@ class ModernModePlugin {
           .filter(a => a.tagName === 'script' && a.attributes)
         legacyAssets.forEach(a => { a.attributes.nomodule = '' })
 
-        if (needsSafariFix) {
-          if (this.unsafeInline) {
-            // inject inline Safari 10 nomodule fix
-            tags.push({
-              tagName: 'script',
-              closeTag: true,
-              innerHTML: safariFix
-            })
-          } else {
-            // inject the fix as an external script
-            const safariFixPath = path.join(this.jsDirectory, 'safari-nomodule-fix.js')
-            const fullSafariFixPath = path.join(compilation.options.output.publicPath, safariFixPath)
-            compilation.assets[safariFixPath] = {
-              source: function () {
-                return Buffer.from(safariFix)
-              },
-              size: function () {
-                return Buffer.byteLength(safariFix)
-              }
-            }
-            tags.push({
-              tagName: 'script',
-              closeTag: true,
-              attributes: {
-                src: fullSafariFixPath
-              }
-            })
-          }
-        }
-
         tags.push(...legacyAssets)
         await fs.remove(tempFilename)
         cb()
@@ -129,5 +83,4 @@ class ModernModePlugin {
   }
 }
 
-ModernModePlugin.safariFix = safariFix
 module.exports = ModernModePlugin
